refactor(cv): extract CvBody from PortfolioGalleryTheme for variant reuse (#45)
The two-column body (Experience, Education, Certifications + Skills, Languages, Soft Skills, Interests, References) was inlined in PortfolioGalleryTheme. Extracted into src/components/cv/themes/shared/cv-body.tsx so every hero variant can render its bespoke hero above a shared body, which is what we will compare across the 6 preview variants.

PortfolioGalleryTheme now renders only its hero and delegates to <CvBody data={data} locale={locale} />. No visual change to /cv — only internal refactor.
